parser improvements
