+++ embl-api-core/src/main/java/uk/ac/ebi/embl/api/validation/helper/Utils.java
@@ -45,7 +45,12 @@ public class Utils {
 	
 	private static final DataManager dataManager = new CachedFileDataManager();
 	private static final CheckFileManager tsvFileManager = new CheckFileManager();
-	
+	private static final Pattern entityNumberPattern = Pattern.compile("^&#([0-9]{2,3});$");
+	private static final Map<String, Character> entityNameToChar = new HashMap<String, Character>() {{
+		put("&amp;", '&');
+		put("&lt;", '<');
+		put("&gt;", '>');
+	}};
 	
 	//prevent instantiation
 	private Utils() {
@@ -965,5 +970,52 @@ public class Utils {
 		}
 		return true;
 	}
-   
+
+	public static StringBuilder escapeASCIIHtmlEntities(final CharSequence input) {
+		if(input == null)
+			return null;
+
+		StringBuilder replaced = new StringBuilder();
+		StringBuilder toBeReplaced = new StringBuilder();
+		boolean isEntity = false;
+
+		for(int i =0; i<input.length();i++){
+			char c = input.charAt(i);
+			if (isEntity) {
+				if (c == '&') {
+					replaced.append(toBeReplaced);
+					toBeReplaced = new StringBuilder();
+					toBeReplaced.append(c);
+				} else {
+					toBeReplaced.append(c);
+					if (c == ';') {
+						if (entityNameToChar.containsKey(toBeReplaced.toString())) {
+							replaced.append(entityNameToChar.get(toBeReplaced.toString()));
+						} else {
+							Matcher m = entityNumberPattern.matcher(toBeReplaced);
+							if (m.matches()) {
+								int ascii = Integer.parseInt(m.group(1));
+								replaced.append(ascii >= 32 && ascii <= 126 ? Character.toString((char) ascii) : toBeReplaced);
+							} else {
+								replaced.append(toBeReplaced);
+							}
+						}
+						isEntity = false;
+						toBeReplaced = new StringBuilder();
+					}
+				}
+			} else if (c == '&') {
+				isEntity = true;
+				toBeReplaced.append(c);
+			} else {
+				replaced.append(c);
+			}
+		}
+
+		if (isEntity) {
+			replaced.append(toBeReplaced);
+		}
+
+		return replaced;
+	}
 }
