+++ embl-api-ff/src/test/java/uk/ac/ebi/embl/flatfile/reader/embl/RAReaderTest.java
@@ -26,18 +26,20 @@ public class RAReaderTest extends EmblReaderTest {
 	
 	public void testRead_Authors1() throws IOException {
 		initLineReader(
-        		"RA   Puzio P.A., von Blau A., Ebneth M.;\n" 
+        		"RA   Puzio P.A., von Blau A., Ebneth M., Cook.A;\n"
 		);		
 		Publication publication = lineReader.getCache().getPublication();
 		ValidationResult result = (new RAReader(lineReader)).read(entry);
 		assertEquals(0, result.count(Severity.ERROR));
-		assertEquals(3, publication.getAuthors().size());
+		assertEquals(4, publication.getAuthors().size());
 		assertEquals("Puzio", publication.getAuthors().get(0).getSurname());
 		assertEquals("P.A.",publication.getAuthors().get(0).getFirstName());
 		assertEquals("von Blau", publication.getAuthors().get(1).getSurname());
 		assertEquals("A.",publication.getAuthors().get(1).getFirstName());
 		assertEquals("Ebneth", publication.getAuthors().get(2).getSurname());
 		assertEquals("M.",publication.getAuthors().get(2).getFirstName());
+		assertEquals("Cook.A",publication.getAuthors().get(3).getFirstName());
+		assertEquals(null,publication.getAuthors().get(3).getSurname());
 	}	
 	
 	public void testRead_Authors2() throws IOException {
@@ -73,7 +75,7 @@ public class RAReaderTest extends EmblReaderTest {
 		assertEquals("J.",publication.getAuthors().get(10).getFirstName());		
 	}
 
-	public void testRead_InvalidAuthors() throws IOException {
+/*	public void testRead_InvalidAuthors() throws IOException {
 		initLineReader(
     		"RA   A. B. C.\n"
 		);
@@ -81,7 +83,7 @@ public class RAReaderTest extends EmblReaderTest {
 		ValidationResult result = (new RAReader(lineReader)).read(entry);
 		assertEquals(1, result.count("RA.1", Severity.ERROR));
 		assertEquals(0, publication.getAuthors().size());
-	}
+	}*/
 
 	public void testRead_NoAuthors() throws IOException {
 		initLineReader(
